refactor[#226] : 단순히 모달을 여는 역할을 하는 컴포넌트는 page 레이어에서 정의
