+++ src/features/setting/ui/index.ts
@@ -1,2 +1,1 @@
-export * from "./Account";
 export * from "./setting";
